requirements.txt
aiogram==3.1.1
pydantic-settings==2.0.3
python-dateutil==2.8.2
motor==3.3.1